requirements.txt
flask==3.0.2
duckdb==0.9.2
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
